refactor: JwtAuthenticationFilter의 authentication 생성, 설정 관련 책임을 AuthenticationUtil 로 위임
