+++ src/main/java/com/goat/server/global/util/filter/JwtAuthenticationFilter.java
@@ -1,55 +1,31 @@
 package com.goat.server.global.util.filter;
 
-import com.goat.server.global.util.jwt.JwtTokenProvider;
+import com.goat.server.global.util.AuthenticationUtil;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
 
-import static com.goat.server.global.domain.type.JwtValidationType.VALID_JWT;
 
 @Component
 @RequiredArgsConstructor
 public class JwtAuthenticationFilter extends OncePerRequestFilter {
 
-    private final JwtTokenProvider jwtTokenProvider;
+    private final AuthenticationUtil authenticationUtil;
 
     @Override
     protected void doFilterInternal(HttpServletRequest request,
                                     HttpServletResponse response,
                                     FilterChain filterChain) throws ServletException, IOException {
 
-
-        final String token = getJwtFromRequest(request);
-
-        if (token != null && jwtTokenProvider.validateToken(token) == VALID_JWT) {
-
-            UserAuthentication authentication = UserAuthentication.from(jwtTokenProvider.getJwtUserDetails(token));
-
-            authentication.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-
-            SecurityContextHolder.getContext().setAuthentication(authentication);
-        }
+        authenticationUtil.setAuthenticationFromRequest(request);
 
         filterChain.doFilter(request, response);
-    }
-
-    private String getJwtFromRequest(HttpServletRequest request) {
-        String bearerToken = request.getHeader("Authorization");
-
-        if (StringUtils.hasText(bearerToken) && bearerToken.startsWith("Bearer ")) {
-            return bearerToken.substring("Bearer ".length());
-        }
-
-        return null;
     }
 
 }
