+++ src/test/java/com/goat/server/auth/presentation/AuthControllerTest.java
@@ -5,7 +5,7 @@ import com.goat.server.auth.application.KakaoSocialService;
 import com.goat.server.auth.dto.response.ReIssueSuccessResponse;
 import com.goat.server.auth.dto.response.SignUpSuccessResponse;
 import com.goat.server.global.CommonControllerTest;
-import com.goat.server.global.util.Tokens;
+import com.goat.server.global.util.jwt.Tokens;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
